Move CoroutineWaiter to ntsa_coroutine.h
